# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arletta Gym API",
        "description": "REST API for Arletta Gym Landing Page — Authentication, Memberships, Trainers, Schedules, Bookings, Payments, Signatures",
        "contact": {
            "name": "Arletta Gym Developer",
            "email": "admin@arlettagym.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get user's bookings",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "confirmed",
                                "cancelled",
                                "completed",
                                "no-show"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bookings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Book a class",
                "operationId": "43e6dcca842886074084b8db86af7802",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "schedule_id"
                                ],
                                "properties": {
                                    "schedule_id": {
                                        "type": "integer"
                                    },
                                    "booking_type": {
                                        "type": "string",
                                        "enum": [
                                            "in-person",
                                            "online"
                                        ],
                                        "example": "in-person"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    },
                    "409": {
                        "description": "Class is full or already booked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{id}/cancel": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Cancel a booking (double verification)",
                "operationId": "9d8eeb663633d4d68870aa36255983b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "verification"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    },
                                    "verification": {
                                        "description": "Must type 'CANCEL' to confirm",
                                        "type": "string",
                                        "example": "CANCEL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking cancelled"
                    },
                    "422": {
                        "description": "Verification failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/bookings/{id}/reschedule": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Reschedule a booking (PT or user)",
                "operationId": "dceb798d86838ccabf4de248349a5326",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_schedule_id",
                                    "reason"
                                ],
                                "properties": {
                                    "new_schedule_id": {
                                        "type": "integer"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking rescheduled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/memberships": {
            "get": {
                "tags": [
                    "Memberships"
                ],
                "summary": "List all membership plans",
                "operationId": "020e3b1359f517bbfe750f7efd9e0e7b",
                "parameters": [
                    {
                        "name": "gym_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Memberships retrieved"
                    }
                }
            }
        },
        "/api/memberships/{id}": {
            "get": {
                "tags": [
                    "Memberships"
                ],
                "summary": "Get membership plan detail",
                "operationId": "20102695f2bf440f8a7fe8ae28e9ef7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membership detail retrieved"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}